==> notebooks/step1_hcbb_read_data.ipynb <==
## Read in datasets for 2015, 2016, 2017 and pickle

 - keep specific columns as discussed in brainstorming
 - add column for year

Columns to keep

- National Provider Identifier

- Last Name/Organization Name of the Provider

- Entity Type of the Provider

- City of the Provider

- Zip Code of the Provider

- State Code of the Provider

- Provider Type

- Place of Service

- HCPCS Code

- HCPCS Description

- Number of Services

- Number of Medicare Beneficiaries

- Number of Distinct Medicare Beneficiary/Per Day Services

- Average Medicare Allowed Amount

Data links:

- https://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data/Physician-and-Other-Supplier2017

- https://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data/Physician-and-Other-Supplier2016

- https://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data/Physician-and-Other-Supplier2015

import pandas as pd
import pickle
import numpy as np

## Read in full csv without chunk method

%%time
# read in without using chunk method - preferable since we're not filtering data on the way in
# keep only columns we want to use
# change column headers to change spaces and special characters 

cols = ['National Provider Identifier',
        'Last Name/Organization Name of the Provider',
        'Entity Type of the Provider',
        'City of the Provider',
        'Zip Code of the Provider',
        'State Code of the Provider',
        'Provider Type',
        'Place of Service', 
        'HCPCS Code', 
        'HCPCS Description', 
        'Number of Services',
        'Number of Medicare Beneficiaries',
        'Number of Distinct Medicare Beneficiary/Per Day Services',
        'Average Medicare Allowed Amount']

df_payments_2017 = pd.read_csv('../data/Medicare_Provider_Utilization_and_Payment_Data__Physician_and_Other_Supplier_PUF_CY2017.csv', 
                               usecols = cols, low_memory = False)
df_payments_2017['year'] = 2017
df_payments_2017.columns = df_payments_2017.columns.str.replace(" ", "_").str.replace("/", "_").str.lower()
df_payments_2017.head()


print(df_payments_2017.shape)
df_payments_2017.info(verbose = True, null_counts = True)

#null values in last_name

df_payments_2017.to_pickle("../data/df_payments_2017.pkl")

df_payments_2017 = pd.read_pickle("../data/df_payments_2017.pkl")

df_payments_2017.info(verbose = True, null_counts = True)


# Read in, drop columns, rename columns, and pickle 2016 and 2015 csv files

%%time

df_payments_2016 = pd.read_csv('../data/Medicare_Provider_Utilization_and_Payment_Data__Physician_and_Other_Supplier_PUF_CY2016.csv', 
                               usecols = cols, low_memory = False)
df_payments_2016['year'] = 2016
df_payments_2016.columns = df_payments_2016.columns.str.replace(" ", "_").str.replace("/", "_").str.lower()
df_payments_2016.head()


print(df_payments_2016.shape)
df_payments_2016.info(verbose = True, null_counts = True)

#null values in last_name (136), zip_code (2)

df_payments_2016.to_pickle('../data/df_payments_2016.pkl')

df_payments_2016 = pd.read_pickle('../data/df_payments_2016.pkl')

df_payments_2016.info(verbose = True, null_counts = True)


%%time

df_payments_2015 = pd.read_csv('../data/Medicare_Provider_Utilization_and_Payment_Data__Physician_and_Other_Supplier_PUF_CY2015.csv', 
                               usecols = cols, low_memory = False)
df_payments_2015['year'] = 2015
df_payments_2015.columns = df_payments_2015.columns.str.replace(" ", "_").str.replace("/", "_").str.lower()
df_payments_2015.head()

print(df_payments_2015.shape)
df_payments_2015.info(verbose = True, null_counts = True)

#null values in last_name (145), entity (1), city (4), 1 each in: zip_code, state, provider_type, place, 
# number_of_services, number_of_medicare_beneficiaries, number_of_distinct, average_medicare_allowed

df_payments_2015.to_pickle('../data/df_payments_2015.pkl')

df_payments_2015 = pd.read_pickle('../data/df_payments_2015.pkl')

df_payments_2015.info(verbose = True, null_counts = True)


## Read in using chunks - ended up not using this method as it took longer than just reading in the csv files in one go.

#%%time

#add year column inside for loop

#cols = ['National Provider Identifier', 
#        'Last Name/Organization Name of the Provider', 
#        'Entity Type of the Provider', 
#        'City of the Provider', 
#        'Zip Code of the Provider', 
#        'State Code of the Provider', 
#        'Provider Type', 
#        'Place of Service', 
#        'Number of Services', 
#        'Number of Medicare Beneficiaries', 
#        'Number of Distinct Medicare Beneficiary/Per Day Services', 
#        'Average Medicare Allowed Amount']

#payment_rows = []
#for chunk in pd.read_csv('../data/Medicare_Provider_Utilization_and_Payment_Data__Physician_and_Other_Supplier_PUF_CY2017.csv', 
#                         chunksize = 1000, usecols = cols):
#    chunk['year'] = 2017
#    payment_rows.append(chunk) 
                
#df_payments_2017 = pd.concat(payment_rows, ignore_index=True)

#CPU times: user 1min 10s, sys: 21 s, total: 1min 31s
#Wall time: 1min 32s


#%%time
#add year column outside for loop, more efficient

#cols = ['National Provider Identifier',
#        'Last Name/Organization Name of the Provider',
#        'Entity Type of the Provider',
#        'City of the Provider',
#        'Zip Code of the Provider',
#        'State Code of the Provider',
#        'Provider Type',
#        'Place of Service',
#        'Number of Services',
#        'Number of Medicare Beneficiaries',
#        'Number of Distinct Medicare Beneficiary/Per Day Services',
#        'Average Medicare Allowed Amount']

#payment_rows = []
#for chunk in pd.read_csv('../data/Medicare_Provider_Utilization_and_Payment_Data__Physician_and_Other_Supplier_PUF_CY2017.csv', 
#                         chunksize = 1000, usecols = cols):
#    payment_rows.append(chunk)
    
#df_payments_2017 = pd.concat(payment_rows, ignore_index=True)
#df_payments_2017['year'] = 2017

#CPU times: user 1min 5s, sys: 23.7 s, total: 1min 29s
#Wall time: 1min 29s
