==> notebooks/step2_medicare_data_cleaning.ipynb <==
## Medicare data cleaning

import pandas as pd
import pickle
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
from glob import glob

%matplotlib inline

# Read in pickle files for 2015, 2016, 2017 and combine

df_payments_2015 = pd.read_pickle('../data/df_payments_2015.pkl')
df_payments_2016 = pd.read_pickle('../data/df_payments_2016.pkl')
df_payments_2017 = pd.read_pickle("../data/df_payments_2017.pkl")

files = sorted(glob('../data/df_payments_*.pkl'))

df_payments_combined = pd.concat((pd.read_pickle(file)
          for file in files), ignore_index= True)

df_payments_combined.head()

df_payments_combined.shape

We looked at the row that had no data except NPI and provider name in the 2015 data - it's copyright info but not actual data so we removed it from the data.

df_payments_combined[df_payments_combined.national_provider_identifier == 1]

df_payments_combined = df_payments_combined.drop(labels = 7205022)

df_payments_combined[df_payments_combined.national_provider_identifier == 1]

# Add column to indicate type of payment - doctor only, facility only, or doctor and facility combined

Entity Type
- I = Individual
- O = Organization

Place of Service
- O = Office
- F = Facility

Entity, Place = Description
- I, F = Doctor Only
- I, O = Doctor & Facility O, F = Facility Only
- O, O = Doctor & Facility

# Step 1 - add new column with payment type (Code from Deigo)

conditions = [
    (df_payments_combined.place_of_service == 'O'),
    (df_payments_combined.entity_type_of_the_provider == 'I') & 
    (df_payments_combined.place_of_service == 'F'),
    (df_payments_combined.entity_type_of_the_provider == 'O') & 
    (df_payments_combined.place_of_service == 'F')
]

choices = ['Doctor & Facility', 'Doctor Only', 'Facility Only']

# Step 2 - add new column with payment type (Code from Deigo)

df_payments_combined['payment_type'] = np.select(conditions, choices, default = 'unknown')

# confirm there are no "unknown" values

df_payments_combined['payment_type'].unique()

# confirm that data looks right

df_payments_combined.tail()

df_payments_combined.to_pickle('../data/payments_combined.pkl')
