==> notebooks/step3_eda.ipynb <==
# EDA on Medicare data 2015-2017

import pandas as pd
import pickle
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

%matplotlib inline

# Read in Combined DataFrame

This combined df has already eliminted the one irrelevant row from 2015, and has added a column for payment_type (Doctor Only, Facility Only, Doctor & Facility).

df_payments_combined = pd.read_pickle('../data/payments_combined.pkl')
print(df_payments_combined.shape)
df_payments_combined.head()

# EDA of combined data

df_payments_combined.shape

df_payments_combined.columns

df_payments_combined.head()

df_payments_combined.info(verbose = True, null_counts = True)

# look at null values

df_payments_combined.isnull().sum()

# 427 null values for provider name. When we looked at this in the data visualizer on the Medicare website,
# these rows had first names (a column we chose to drop from our dataset) and all other pertinent values.
# Since the NPI is unique to each provider, we determined that a null value in the provider name column was not
# an impediment to analysis.

# number of unique HCPCS codes used

count_of_codes = df_payments_combined.hcpcs_code.nunique()
print("There are", count_of_codes, "unique HCPCS codes in 2015-2017")

# number_of_services stats 
# min 2.4 (how is this not a round number?)
# max 7,195,536

# note: mean is 246 and median is 44

df_payments_combined.number_of_services.describe()

df_payments_combined.number_of_medicare_beneficiaries.describe()

# mean 143.7782
# median 41

df_payments_combined.number_of_distinct_medicare_beneficiary_per_day_services.describe()

# mean 100.43
# median 63.99

# min 0.0000603 (check, this doesn't seem valid, although the numbers in this column are averages for a provider
# for a single HCPCS code so maybe it is)

df_payments_combined.average_medicare_allowed_amount.describe()

Significant difference between mean and median in the stats for the measures above.

# Creating new df showing average payment by year, then by payment type, then by HCPCS code. (Code from Nicole)

df_avg_pmt = df_payments_combined.groupby(['year',
                                           'payment_type',
                                           'hcpcs_code']).average_medicare_allowed_amount.mean().to_frame()

df_avg_pmt = df_avg_pmt.reset_index()
df_avg_pmt.head()

df_avg_pmt.shape

df_avg_pmt.info()

df_med_services_day = df_payments_combined.groupby(['year',
                                                    'hcpcs_code']).number_of_distinct_medicare_beneficiary_per_day_services.mean().to_frame().reset_index()

df_med_services_day.head()

# still seeing wide variance between mean and median
# examine outliers

df_avg_pmt.describe()

# still seeing wide variance between mean and median 
# examine outliers

df_med_services_day.describe()

# Pivot dataframe

pivot_index = ['national_provider_identifier',
               'entity_type_of_the_provider', 
               'place_of_service',
               'payment_type',
               'provider_type',
               'hcpcs_code',
               'hcpcs_description',
               'zip_code_of_the_provider', 
               'state_code_of_the_provider']

pivot_cols = ['year']

%%time
df_pmt_pvt = df_payments_combined.pivot_table(index = pivot_index, 
                                              columns = pivot_cols, 
                                              values = 'average_medicare_allowed_amount', 
                                              aggfunc=np.mean)

df_pmt_pvt = df_pmt_pvt.reset_index()


print(df_pmt_pvt.shape)
df_pmt_pvt.head()

%%time
df_bpd_pvt = df_payments_combined.pivot_table(index = pivot_index, 
                                              columns = pivot_cols, 
                                              values = 'number_of_distinct_medicare_beneficiary_per_day_services', 
                                              aggfunc=np.mean)

df_bpd_pvt = df_bpd_pvt.reset_index()


print(df_bpd_pvt.shape)
df_bpd_pvt.head()

df_pmt_pvt.describe()

df_pmt_pvt.columns

df_pmt_pvt.shape

df_pmt_pvt.to_pickle('../data/df_pmt_pvt.pkl')

df_bpd_pvt.to_pickle('../data/df_bpd_pvt.pkl')

df_pmt_pvt_group = df_pmt_pvt.groupby(['payment_type', 
                                       'hcpcs_code'])

df_pmt_pvt_group.head()

df_pmt_pvt_group.hcpcs_code.value_counts()

df_pmt_pvt_group_diff = df_pmt_pvt_group
df_pmt_pvt_group_diff.head()

# not working - pivot dataframe is wierd about grouping/aggregating

df_pmt_pvt__group_diff['pmt_diff_amt'] = df_pmt_pvt_group_diff[2017] - df_pmt_pvt_group_diff[2015]

df_pmt_pvt.to_csv('../data/df_pmt_pvt.csv')

df_payments_combined.to_csv('../data/df_payments_combined.csv')
